@@ -1,17 +1,21 @@
 {
   "name": "zongphp/alipay",
-  "description": "支付宝组件",
-  "type": "project",
+  "description": "支付宝组件用于实现支付宝在线支付功能",
   "keywords": [
     "alipay",
     "支付宝支付"
   ],
   "license": "MIT",
-  "minimum-stability": "dev",
+  "minimum-stability": "stable",
   "require": {
-    "php": ">=5.4.0"
+    "php": ">=5.6.0",
+    "zongphp/config": "~2.0",
+    "zongphp/request": "~2.0"
   },
   "autoload": {
+    "classmap": [
+      "src/org/"
+    ],
     "files": [],
     "psr-4": {
       "zongphp\\alipay\\": "src"
